+++ feature-service/src/main/java/ba/unsa/etf/pnwt/service/skill/SkillServiceImpl.java
@@ -1,12 +1,18 @@
 package ba.unsa.etf.pnwt.service.skill;
 
 import ba.unsa.etf.pnwt.dto.SkillDTO;
+import ba.unsa.etf.pnwt.entity.SkillEntity;
+import ba.unsa.etf.pnwt.entity.UserEntity;
 import ba.unsa.etf.pnwt.mapper.SkillMapper;
+import ba.unsa.etf.pnwt.mapper.UserMapper;
 import ba.unsa.etf.pnwt.repository.SkillRepository;
+import ba.unsa.etf.pnwt.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class SkillServiceImpl implements SkillService{
@@ -14,9 +20,16 @@ public class SkillServiceImpl implements SkillService{
     @Autowired
     private SkillRepository skillRepository;
 
+    @Autowired
+    private UserRepository userRepository;
+
     @Override
     public List<SkillDTO> findByUserId(Integer id) {
-        return SkillMapper.mapToProjections(skillRepository.findByUserId(id));
+        UserEntity userEntity = userRepository.findById(id);
+        List<SkillEntity> skillEntity = skillRepository.findByUserId(id);
+        if (userEntity != null)
+            return SkillMapper.mapToProjections(skillEntity);
+        else throw new EntityNotFoundException("User with provided id not found");
     }
 
     @Override
@@ -26,8 +39,11 @@ public class SkillServiceImpl implements SkillService{
 
     @Override
     public String deleteSkill(Integer id) {
-        skillRepository.deleteSkill(id);
-        return "Skill was successfully deleted";
+        Optional<SkillEntity> skillEntity = skillRepository.findById(id);
+        if (skillEntity.isPresent()) {
+            skillRepository.deleteSkill(id);
+            return "Skill was successfully deleted";
+        }else throw new EntityNotFoundException("Skill with provided id not found");
     }
 
     @Override
